See details below...
Please add a code example of which value(s)? to save to my local storage and restore such that my ion-list displays in the same position I scrolled last to in the browser.  Thank you.
The following isn't working
//  save
this.scrollPosition = this.content.contentTop
// restore
this.content.scrollTo(0, this.scrollPosition)

Can you please add a diagram to visualize all these properties with seamingly similar definitions.  
Define "content" eg. all the items in the list regardless whether they are visible in the current view or not.
 
contentHeight
Content height of the viewable area. This does not include content which is outside the overflow area, or content area which is under headers and footers. Read-only.
 scrollHeight
Content height including content which is not visible on the screen due to overflow. Read-only